Adding some of my old files. The most important here is truthmakers_fancy.py, which is the fully commented version of my code.
I'm also adding some of the .dot files I was outputting before, and a few pdfs.  See mydot.dot where I put in a grouping into subgraphs by hand, generating output along the lines of what you see in testdot.pdf.  (This 'subgraph' feature in Graphviz is going to be important if we want to generate output that involves some side-by-side Hasse diagrams with connecting lines.)